BIG 5 Persönlichkeitstest: Vollständig

Fragebogen zu den BIG 5 Persönlichkeitsfaktoren durch das 50-Item Lexical Big-Five Factor Markers Inventar (IPIP). Jede der fünf Persönlichkeitseigenschaften wurde mithilfe von jeweils 10 Items gemessen. Die unterschiedlichen Fragen konnten auf einer Likert Skala beantwortet werden (1: Sehr unzutreffend, 2: Moderat unzutreffend, 3: Weder zutreffend noch unzutreffend, 4: Moderat zutreffend, 5: Sehr zutreffend).

Invers kodierte Items sind in der Beschreibung unten entsprechend markiert. Der Anfangsbuchstabe der Spalte drückt den Persönlichkeitsfaktor aus: Extraversion, Neurotizismus, Verträglichkeit, Gewissenhaftigkeit und Offenheit für neue Erfahrungen).

Verwendung

big5_comp

Format

Ein tibble mit 5000 Zeilen and 52 Spalten:

Alter

Alter in Lebensjahren

Geschlecht

f: Frauen, m: Männer

E1

Ich bringe eine Party in Schwung.

E2

Ich spreche nicht viel. (INVERS)

E3

Ich fühle mich wohl unter Menschen.

E4

Ich halte mich im Hintergrund. (INVERS)

E5

Ich beginne Gespräche.

E6

Ich habe wenig zu sagen. (INVERS)

E7

Ich spreche mit vielen verschiedenen Leuten auf Parties.

E8

Ich mag es nicht, Aufmerksamkeit auf mich zu ziehen. (INVERS)

E9

Es macht mir nichts aus, im Zentrum der Aufmerksamkeit zu stehen.

E10

Ich bin still unter Fremden. (INVERS)

N1

Ich gerate leicht in Stress.

N2

Ich bin die meiste Zeit entspannt. (INVERS)

N3

Ich grübel über Dinge.

N4

Ich fühle mich selten deprimiert oder bedrückt. (INVERS)

N5

Ich bin leicht beunruhigt.

N6

Ich rege mich leicht auf.

N7

Ich wechsel oft meine Stimmung.

N8

Ich habe häufig Stimmungsschwankungen.

N9

Ich bin leicht gereizt.

N10

fühle mich oft deprimiert oder bedrückt.

V1

Ich empfinde wenig für andere. (INVERS)

V2

Ich bin interessiert an anderen Menschen.

V3

Ich beleidige andere. (INVERS)

V4

Ich fühle mit anderen Menschen mit.

V5

Ich bin nicht interessiert an den Problemen anderer Leute. (INVERS)

V6

Ich habe ein weiches Herz.

V7

Ich bin nicht wirklich interessiert an anderen. (INVERS)

V8

Ich nehme mir Zeit für andere.

V9

Ich fühle die Gefühle anderer.

V10

Ich kann andere beruhigen.

G1

Ich bin immer vorbereitet.

G2

Ich lasse meine Sachen herumliegen. (INVERS)

G3

Ich achte auf Details.

G4

Ich verursache großes Durcheinander. (INVERS)

G5

Ich erledige Hausarbeiten unmittelbar.

G6

Ich vergesse oft, Dinge an ihren Platz zurückzulegen. (INVERS)

G7

Ich mag Ordnung.

G8

Ich drücke mich vor meinen Pflichten. (INVERS)

G9

Ich folge einem Plan.

G10

Ich bin anspruchsvoll in meiner Arbeit.

O1

Ich besitze einen großen Wortschatz.

O2

Ich habe Schwierigkeiten abstrakte Ideen zu verstehen. (INVERS)

O3

Ich habe eine lebhafte Vorstellungskraft.

O4

Ich bin an abstrakten Ideen nicht interessiert. (INVERS)

O5

Ich habe ausgezeichnete Ideen.

O6

Ich habe kein gutes Vorstellungsvermögen. (INVERS)

O7

Ich verstehe Dinge schnell.

O8

Ich gebrauche schwierige Wörter.

O9

Ich verbringe Zeit damit über Dinge nachzudenken.

O10

Ich bin voller Ideen.

Quelle

https://openpsychometrics.org/_rawdata/

Beispiele

big5_comp
#> # A tibble: 5,000 × 52
#>    Alter Geschlecht    E1    E2    E3    E4    E5    E6    E7    E8    E9   E10
#>    <dbl> <chr>      <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1    62 f              4     2     5     2     5     1     4     3     5     1
#>  2    14 m              2     2     3     3     3     3     1     5     1     5
#>  3    11 m              5     1     1     4     5     1     1     5     5     1
#>  4    43 m              2     5     2     4     3     4     3     4     4     5
#>  5    68 m              3     1     3     3     3     1     3     1     3     5
#>  6    22 m              1     5     2     4     1     3     2     4     1     5
#>  7    42 m              5     1     5     1     5     1     5     4     4     1
#>  8    71 f              4     3     5     3     5     1     4     3     4     3
#>  9    50 m              3     1     5     1     5     1     5     2     5     3
#> 10    34 m              1     4     2     5     2     4     1     4     1     5
#> # ℹ 4,990 more rows
#> # ℹ 40 more variables: N1 <dbl>, N2 <dbl>, N3 <dbl>, N4 <dbl>, N5 <dbl>,
#> #   N6 <dbl>, N7 <dbl>, N8 <dbl>, N9 <dbl>, N10 <dbl>, V1 <dbl>, V2 <dbl>,
#> #   V3 <dbl>, V4 <dbl>, V5 <dbl>, V6 <dbl>, V7 <dbl>, V8 <dbl>, V9 <dbl>,
#> #   V10 <dbl>, G1 <dbl>, G2 <dbl>, G3 <dbl>, G4 <dbl>, G5 <dbl>, G6 <dbl>,
#> #   G7 <dbl>, G8 <dbl>, G9 <dbl>, G10 <dbl>, O1 <dbl>, O2 <dbl>, O3 <dbl>,
#> #   O4 <dbl>, O5 <dbl>, O6 <dbl>, O7 <dbl>, O8 <dbl>, O9 <dbl>, O10 <dbl>