Antwortmatrix erstellen

Erstellt eine binäre Ergebnismatrix, in der die richtigen Antworten mit einer 1 und die falschen mit einer 0 ausgegeben werden. Falls die richtigen Antworten sowohl Numerics als auch Characters beinhalten, kann der Funktion ebenfalls ein einzeiliger data.frame oder tibble übergeben werden. Fehlende Werte werden automatisch als falsch beantwortet gewertet.

Verwendung

data_binary(df, answers)

Argumente

df

Datensatz als data.frame oder tibble.

answers

Antwortvektor als Vektor oder einzeiliger data.frame bzw. tibble.

Rückgabewert

Antwortmatrix als tibble.

Beispiele

# Identische Datentypen
soscisurvey <- data.frame(
 I1 = c(3, 2, 3, 1),
 I2 = c(2, 2, 2, 1),
 I3 = c(4, 1, 4, 4)
)
soscisurvey
#>   I1 I2 I3
#> 1  3  2  4
#> 2  2  2  1
#> 3  3  2  4
#> 4  1  1  4
antwortvec <- c(3, 2, 4)
data_binary(df = soscisurvey, answers = antwortvec)
#> # A tibble: 4 × 3
#>      I1    I2    I3
#>   <dbl> <dbl> <dbl>
#> 1     1     1     1
#> 2     0     1     0
#> 3     1     1     1
#> 4     0     0     1

# Verschiedene Datentypen
soscisurvey2 <- data.frame(
 I1 = c(3, 2, 3, 1),
 I2 = c("Apfel", "Banane", "Erdbeere", "Apfel"),
 I3 = c(4, 1, NA, 4)
)
soscisurvey2
#>   I1       I2 I3
#> 1  3    Apfel  4
#> 2  2   Banane  1
#> 3  3 Erdbeere NA
#> 4  1    Apfel  4
antwortdf <- data.frame(3, "Apfel", 4)
data_binary(df = soscisurvey2, answers = antwortdf)
#> # A tibble: 4 × 3
#>      I1    I2    I3
#>   <dbl> <dbl> <dbl>
#> 1     1     1     1
#> 2     0     0     0
#> 3     1     0     0
#> 4     0     1     1