Statistische Bildung

In dieser Studie mit 146 ProbandInnen wurde zum einen versucht, einen Zusammenhang zwischen statistischer Bildung und Risikoverhalten zu finden und zum anderen eine Verbesserung statistischer Bildung durch eine graphische Hilfestellung zu entdecken.

Dafür wurden zwei verschiedene Fragebogenversionen ausgeteilt. In der einen gab es nach rund der Hälfte der Aufgaben eine graphische Hilfestellung, die eine mögliche Lösungsstrategie zu einem Teil der Aufgaben erklärt (Interventionsgruppe). Die andere Version beinhaltete lediglich eine schriftliche Erklärung. Statistische Bildung (engl. Statistical literacy) ist als Interpretieren, kritisches Bewerten und die Fähigkeit über statistische Informationen zu sprechen definiert.

Verwendung

bildung

Format

Ein tibble mit 146 Zeilen and 71 Spalten:

ID

Versuchspersonennummer

Version

Fragebogenversion mit Intervetion und ohne

Geschlecht

Biologisches Geschlecht mit f:weiblich, m:männlich.

Alter

Alter in Lebensjahren

Schulbildung

Höchste abgeschlossene Schulausbildung.

Risikosportarten

Praktizieren mindestens einer Risikosportart.

Statistische_Kenntnisse

Selbsteinschätzung der eigenen Statistikkenntnisse auf einer Skala von 0 (keine Kenntnisse) bis 6 (ausgezeichnete Kenntnisse).

Interesse_Wissenschaft

Subjektives Interesse an Wissenschaft auf einer Skala von 0 (kein Interesse) bis 6 (sehr starkes Interesse).

Dospert_verhalten

Median der DOSPERT Skala zum Risikoverhalten mit einer Skala von 1 (sehr unwahrscheinlich) bis 7 (sehr wahrscheinlich).

Dospert_wahrnehmung

Median der DOSPERT Skala zur Risikowahrnehmung mit einer Skala von 1 (überhaupt kein Risiko) bis 7 (sehr hohes Risiko).

Dospert_nutzen

Median der DOSPERT Skala zum Risikonutzen mit einer Skala von 1 (gar kein nutzen) bis 7 (sehr hoher Nutzen).

Bildung_pre

Mittlere Lösungswahrscheinlichkeit über alle Fragen vor der Intervention.

Bildung_post

Mittlere Lösungswahrscheinlichkeit über alle Fragen nach Intervention.

Bildung_gesamt

Mittlere Lösungswahrscheinlichkeit insgesamt.

Beispiele

bildung
#> # A tibble: 146 × 14
#>    ID    Fragebogen   Geschlecht Alter Schulbildung        Risikosportarten
#>    <chr> <fct>        <fct>      <dbl> <chr>                          <dbl>
#>  1 1     Intervention f             26 Abitur                             1
#>  2 2     Keine        m             26 Abitur                             0
#>  3 3     Keine        f             28 Hauptschulabschluss                0
#>  4 4     Intervention m             26 Hochschulabschluss                 1
#>  5 5     Intervention m             25 Abitur                             1
#>  6 6     Keine        f             24 Lehrabschluss                      0
#>  7 7     Intervention f             55 Lehrabschluss                      0
#>  8 8     Keine        m             26 Hochschulabschluss                 0
#>  9 9     Intervention m             54 Hochschulabschluss                 1
#> 10 10    Keine        f             53 Hochschulabschluss                 1
#> # ℹ 136 more rows
#> # ℹ 8 more variables: Statistische_kenntnisse <dbl>,
#> #   Interesse_Wissenschaft <dbl>, Dospert_verhalten <dbl>,
#> #   Dospert_wahrnehmung <dbl>, Dospert_nutzen <dbl>, Bildung_pre <dbl>,
#> #   Bildung_post <dbl>, Bildung_gesamt <dbl>