Lagemaße und Streuungsmaße berechnen

Berechnet die Anzahl N, Quartile, Mittelwert, Median, Standardabweichung und Standardfehler.

Verwendung

descriptive(data, .round = 2, .print = NA, .tibble = TRUE)

Argumente

data

Daten als data.frame, tibble oder grouped.df.

.round

Anzahl der Nachkommastellen.

.print

Anzahl der Zeilen, die ausgegeben werden sollen. Bei Angabe einer Zahl wird das Standardverhalten eines tibbles, i.d.R nur 10 Zeilen anzuzeigen, überschrieben.

.tibble

Ausgabe als data.frame (FALSE) oder tibble (TRUE).

Rückgabewert

Überblick über typische Lage- und Streuungsmaße.

Beispiele

chemo
#> # A tibble: 450 × 13
#>    Pat_id Geschlecht Behandlung Leukos_t6 Leukos_t0 Alter Infektionen Stationaer
#>     <int> <chr>      <fct>          <dbl>     <dbl> <dbl>       <int>      <int>
#>  1      1 f          Radio           5.68      5.74    49           4          0
#>  2      2 f          Radiochemo      6.64      6.71    46           3          0
#>  3      3 m          Radio           5.97      6.03    41           3          0
#>  4      4 f          Radio           5.55      5.6     43           1          1
#>  5      5 f          Radio           5.6       5.66    47           3          0
#>  6      6 f          Chemo           5.96      6.02    45           4          0
#>  7      7 m          Chemo           5.6       5.65    50           4          0
#>  8      8 m          Chemo           5.4       5.45    43           3          0
#>  9      9 m          Radiochemo      5.92      5.98    37           1          0
#> 10     10 m          Radiochemo      5.63      5.68    47           2          0
#> # ℹ 440 more rows
#> # ℹ 5 more variables: Komorb <fct>, Beob_zeit <dbl>, Status <dbl>,
#> #   Schmerzen <int>, Lebensqualitaet <int>

chemo |> 
  descriptive()
#> # A tibble: 10 × 10
#>    Variable            N   Min     Q1   Mean Median     Q3    Max     SD    SE
#>    <chr>           <int> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl> <dbl>
#>  1 Alter             450 33     42     44.8   45     48     57      4.31  0.2 
#>  2 Beob_zeit         450  0.07   0.99   2.4    1.83   3.25  15.0    2.05  0.1 
#>  3 Infektionen       450  0      1      2.26   2      3     10      2.05  0.1 
#>  4 Lebensqualitaet   450  1      2      3.59   3      5     10      2.2   0.1 
#>  5 Leukos_t0         450  5.1    5.8    6.04   6.05   6.27   6.92   0.33  0.02
#>  6 Leukos_t6         450  5.05   5.74   5.98   5.99   6.21   6.85   0.33  0.02
#>  7 Pat_id            450  1    113.   226.   226.   338.   450    130.    6.13
#>  8 Schmerzen         450  1      3      5.15   5      7     10      2.35  0.11
#>  9 Stationaer        450  0      0      0.09   0      0      1      0.28  0.01
#> 10 Status            450  0      1      0.9    1      1      1      0.29  0.01

chemo |> 
  dplyr::group_by(Behandlung) |> 
  descriptive(.print = 25)
#> # A tibble: 30 × 11
#> # Groups:   Behandlung [3]
#>    Behandlung Variable         N   Min     Q1   Mean Median     Q3    Max     SD
#>    <fct>      <chr>        <int> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 Radiochemo Alter          151 35     41.5   44.7   45     47.5   56      4.41
#>  2 Radiochemo Beob_zeit      151  0.18   2.04   3.84   3.31   5.06  15.0    2.58
#>  3 Radiochemo Infektionen    151  0      1      1.98   1      3     10      2.05
#>  4 Radiochemo Lebensquali…   151  1      2      3.39   3      5     10      2.2 
#>  5 Radiochemo Leukos_t0      151  5.1    5.88   6.08   6.07   6.32   6.9    0.35
#>  6 Radiochemo Leukos_t6      151  5.05   5.82   6.02   6.01   6.26   6.83   0.35
#>  7 Radiochemo Pat_id         151  2    106.   220.   218    330.   446    130.  
#>  8 Radiochemo Schmerzen      151  1      3      5.3    5      7     10      2.44
#>  9 Radiochemo Stationaer     151  0      0      0.05   0      0      1      0.21
#> 10 Radiochemo Status         151  0      1      0.87   1      1      1      0.34
#> 11 Chemo      Alter          150 35     42     45.1   45     48     57      4.4 
#> 12 Chemo      Beob_zeit      150  0.07   1.06   2.09   1.8    2.87   6.14   1.35
#> 13 Chemo      Infektionen    150  0      1      2.33   2      3      9      1.97
#> 14 Chemo      Lebensquali…   150  1      2      3.63   3      5     10      2.21
#> 15 Chemo      Leukos_t0      150  5.38   5.81   6.05   6.08   6.29   6.92   0.33
#> 16 Chemo      Leukos_t6      150  5.33   5.75   5.99   6.02   6.23   6.85   0.33
#> 17 Chemo      Pat_id         150  6    121.   231.   238.   342    450    133.  
#> 18 Chemo      Schmerzen      150  1      3      5.14   5      7     10      2.32
#> 19 Chemo      Stationaer     150  0      0      0.07   0      0      1      0.26
#> 20 Chemo      Status         150  0      1      0.91   1      1      1      0.29
#> 21 Radio      Alter          149 33     42     44.5   44     47     55      4.12
#> 22 Radio      Beob_zeit      149  0.14   0.68   1.24   1.05   1.7    3.42   0.73
#> 23 Radio      Infektionen    149  0      1      2.49   2      3     10      2.13
#> 24 Radio      Lebensquali…   149  1      2      3.76   4      5     10      2.19
#> 25 Radio      Leukos_t0      149  5.22   5.74   5.98   6      6.21   6.72   0.31
#> # ℹ 5 more rows
#> # ℹ 1 more variable: SE <dbl>